+++ libs/ui-lib/lib/common/validationSchemas/addressValidation.tsx
@@ -17,7 +17,7 @@ export const ipValidationSchema = (t: TFunction) =>
   Yup.string().test(
     'ip-validation',
     t('ai:Not a valid IP address'),
-    (value?: string) => Address4.isValid(value || '') || Address6.isValid(value || ''),
+    (value?: string) => !value || Address4.isValid(value) || Address6.isValid(value),
   );
 
 export const ipNoSuffixValidationSchema = (t: TFunction) =>
