+++ libs/ui-lib/lib/ocm/components/wizard/steps/disconnected/DisconnectedStaticIp.tsx
@@ -1,30 +1,11 @@
 import React from 'react';
-import {
-  ClusterWizardStep,
-  WithErrorBoundary,
-  getFormikErrorFields,
-  useAlerts,
-  InfraEnvsAPI,
-} from '../../../../../common';
 import { InfraEnvUpdateParams } from '@openshift-assisted/types/assisted-installer-service';
+import { InfraEnvsAPI } from '../../../../../common';
 import { useClusterWizardContext } from '../../clusterWizardContext';
-import { ClusterWizardNavigation, ClusterWizardFooter } from '../../wizardComponents';
-import { StaticIpFormState, StaticIpPage } from '../staticIp/components';
-
-const getInitialFormState = (): StaticIpFormState => ({
-  isValid: true,
-  isSubmitting: false,
-  isAutoSaveRunning: false,
-  errors: {},
-  touched: {},
-  isEmpty: true,
-});
+import { StaticIp } from '../staticIp/StaticIp';
 
 export const DisconnectedStaticIp: React.FC = () => {
-  const { moveNext, moveBack, disconnectedInfraEnv, setDisconnectedInfraEnv } =
-    useClusterWizardContext();
-  const { alerts } = useAlerts();
-  const [formState, setFormState] = React.useState<StaticIpFormState>(getInitialFormState());
+  const { disconnectedInfraEnv, setDisconnectedInfraEnv } = useClusterWizardContext();
 
   const updateInfraEnv = React.useCallback(
     async (params: InfraEnvUpdateParams) => {
@@ -38,37 +19,11 @@ export const DisconnectedStaticIp: React.FC = () => {
     [disconnectedInfraEnv, setDisconnectedInfraEnv],
   );
 
-  const isNextDisabled = !formState.isValid || !!alerts.length || formState.isSubmitting;
-  const errorFields = getFormikErrorFields<object>(formState.errors, formState.touched);
-
-  const footer = (
-    <ClusterWizardFooter
-      alertTitle="Static IP configuration contains missing or invalid fields"
-      alertContent={null}
-      errorFields={errorFields}
-      isSubmitting={formState.isSubmitting}
-      onNext={moveNext}
-      onBack={moveBack}
-      isNextDisabled={isNextDisabled}
-      isBackDisabled={formState.isSubmitting}
-    />
-  );
-
   if (!disconnectedInfraEnv) {
     return null;
   }
 
-  return (
-    <ClusterWizardStep navigation={<ClusterWizardNavigation />} footer={footer}>
-      <WithErrorBoundary title="Failed to load static IP step">
-        <StaticIpPage
-          infraEnv={disconnectedInfraEnv}
-          updateInfraEnv={updateInfraEnv}
-          onFormStateChange={setFormState}
-        />
-      </WithErrorBoundary>
-    </ClusterWizardStep>
-  );
+  return <StaticIp infraEnv={disconnectedInfraEnv} updateInfraEnv={updateInfraEnv} />;
 };
 
 export default DisconnectedStaticIp;
