+++ libs/ui-lib/lib/ocm/components/wizard/steps/staticIp/StaticIp.tsx
@@ -25,7 +25,7 @@ const getInitialFormStateProps = () => {
   };
 };
 
-export const StaticIp: React.FC<StaticIpProps & { cluster: Cluster }> = ({
+export const StaticIp: React.FC<StaticIpProps & { cluster?: Cluster }> = ({
   cluster,
   infraEnv,
   updateInfraEnv,
